simplifying controllers by moving all views into a standalone controller, and leaving the remaining controllers to cover the apis
