--- appsensor-ui/src/main/java/org/owasp/appsensor/ui/controller/GeoMapController.java
@@ -1,15 +1,0 @@
-package org.owasp.appsensor.ui.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-@Controller
-public class GeoMapController {
-
-	@RequestMapping(value="/geo-map", method = RequestMethod.GET)
-	public String home() {
-		return "geo-map";
-	}
-
-}
