+++ appsensor-ui/src/main/java/org/owasp/appsensor/ui/controller/UserController.java
@@ -1,15 +1,10 @@
 package org.owasp.appsensor.ui.controller;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
 public class UserController {
 
-	@RequestMapping(value="/user", method = RequestMethod.GET)
-	public String home() {
-		return "user";
-	}
+	
 
 }
